update options to match what is upstream
Some of the original options matched up with tickets that were modified
later in the process.  Change them to match what we have functionality
for and support for.  Now that all the other features are up this should
be more accurate.  Also went through and verified that the server and
local versions were being called with the correct prompts for them.